--- a/src/cuda_xengine.cu
+++ b/src/cuda_xengine.cu
@@ -77,8 +77,8 @@ static texture<float2, 2, cudaReadModeElementType> tex2dfloat2;
 #else
 #ifdef DP4A
 // texture declaration for swizzled 8-bit fixed point reads
-static texture<char4, 1, cudaReadModeElementType> tex1dchar4;
-static texture<char4, 2, cudaReadModeElementType> tex2dchar4;
+static texture<int, 1, cudaReadModeElementType> tex1dchar4;
+static texture<int, 2, cudaReadModeElementType> tex2dchar4;
 #else
 // texture declaration for 8-bit fixed point reads
 static texture<char2, 1, cudaReadModeNormalizedFloat> tex1dfloat2;
@@ -273,7 +273,7 @@ int xgpuInit(XGPUContext *context, int device_flags)
   internal->channelDesc = cudaCreateChannelDesc<float2>();
 #else
 #ifdef DP4A
-  internal->channelDesc = cudaCreateChannelDesc<char4>();
+  internal->channelDesc = cudaCreateChannelDesc<int>();
 #else
   internal->channelDesc = cudaCreateChannelDesc<char2>();
 #endif // DP4A
@@ -299,18 +299,29 @@ int xgpuInit(XGPUContext *context, int device_flags)
 
   // check whether texture dimensions are ok
 #if TEXTURE_DIM == 2
-  if((NFREQUENCY * NSTATION * NPOL > deviceProp.maxTexture2D[0]) ||
-     (NTIME_PIPE > deviceProp.maxTexture2D[1])) {
+#ifdef DP4A
+  if((NFREQUENCY * NSTATION * NPOL > deviceProp.maxTexture2DLinear[0]) ||
+     (NTIME_PIPE/2 > deviceProp.maxTexture2DLinear[1])) {
     return XGPU_INSUFFICIENT_TEXTURE_MEMORY;
   }
+#else
+  if((NFREQUENCY * NSTATION * NPOL > deviceProp.maxTexture2DLinear[0]) ||
+     (NTIME_PIPE > deviceProp.maxTexture2DLinear[1])) {
+    return XGPU_INSUFFICIENT_TEXTURE_MEMORY;
+  }
+#endif
 #elif TEXTURE_DIM == 1
-#if 0
   // Surprisingly, this appears not to be a problem with 1D textures.  On a
   // GeForce GTX 580 (i.e. Fermi device), deviceQuery returns 65536 as
   // maxTexture1D, yet the default sizes use 10 * 256 * 2 * 100 * 2 == 1024000
   // bytes of 1D texture without any problems.  Perhaps the value of
   // maxTexture1D returned by cudaGetDeviceProperties is wrong?
-  if (NFREQUENCY * NSTATION * NPOL * NTIME_PIPE > deviceProp.maxTexture1D) {
+#ifdef DP4A
+  if (NFREQUENCY * NSTATION * NPOL * NTIME_PIPE/2 > deviceProp.maxTexture1DLinear) {
+    return XGPU_INSUFFICIENT_TEXTURE_MEMORY;
+  }
+#else
+  if (NFREQUENCY * NSTATION * NPOL * NTIME_PIPE > deviceProp.maxTexture1DLinear) {
     return XGPU_INSUFFICIENT_TEXTURE_MEMORY;
   }
 #endif
@@ -591,8 +602,13 @@ int xgpuCudaXengine(XGPUContext *context, int syncOp)
 
     // Kernel Calculation
 #if TEXTURE_DIM == 2
-    cudaBindTexture2D(0, tex2dfloat2, array_compute, channelDesc, NFREQUENCY*NSTATION*NPOL, NTIME_PIPE, 
+#ifndef DP4A
+    cudaBindTexture2D(0, tex2dfloat2, array_compute, channelDesc, NFREQUENCY*NSTATION*NPOL, NTIME_PIPE,
 		      NFREQUENCY*NSTATION*NPOL*sizeof(ComplexInput));
+#else
+    cudaBindTexture2D(0, tex2dchar4, array_compute, channelDesc, NFREQUENCY*NSTATION*NPOL, 2*(NTIME_PIPE/4),
+		      NFREQUENCY*NSTATION*NPOL*sizeof(char4));
+#endif
 #else
 #ifndef DP4A
     cudaBindTexture(0, tex1dfloat2, array_compute, channelDesc, NFREQUENCY*NSTATION*NPOL*NTIME_PIPE*sizeof(ComplexInput));
@@ -618,8 +634,13 @@ int xgpuCudaXengine(XGPUContext *context, int syncOp)
   array_compute = array_d[(PIPE_LENGTH+1)%2];
   // Final kernel calculation
 #if TEXTURE_DIM == 2
-    cudaBindTexture2D(0, tex2dfloat2, array_compute, channelDesc, NFREQUENCY*NSTATION*NPOL, NTIME_PIPE, 
+#ifndef DP4A
+    cudaBindTexture2D(0, tex2dfloat2, array_compute, channelDesc, NFREQUENCY*NSTATION*NPOL, NTIME_PIPE,
 		      NFREQUENCY*NSTATION*NPOL*sizeof(ComplexInput));
+#else
+    cudaBindTexture2D(0, tex2dchar4, array_compute, channelDesc, NFREQUENCY*NSTATION*NPOL, 2*(NTIME_PIPE/4),
+		      NFREQUENCY*NSTATION*NPOL*sizeof(char4));
+#endif
 #else
 #ifndef DP4A
     cudaBindTexture(0, tex1dfloat2, array_compute, channelDesc, NFREQUENCY*NSTATION*NPOL*NTIME_PIPE*sizeof(ComplexInput));