--- a/src/shared_transfer_4_dp4a.cuh
+++ b/src/shared_transfer_4_dp4a.cuh
@@ -6,16 +6,43 @@
 
 #if TEXTURE_DIM == 1
 
-// Read char4 from global, write char4s to shared memory avoid bank conflict.
+// Read char4 from global, write int to shared memory avoid bank conflict.
 #define LOAD(s, t)							\
-  {char4 real = tex1Dfetch(tex1dchar4, array_index + (t)*NFREQUENCY*NSTATION*NPOL); \
-    char4 imag = tex1Dfetch(tex1dchar4, array_index + (NTIME_PIPE/4 + t)*NFREQUENCY*NSTATION*NPOL); \
-    CUBE_ADD_BYTES(sizeof(ComplexInput));				\
-    *(input##s##_p) = *((int*)&real);					\
-    *(input##s##_p + 4*TILE_WIDTH) = *((int*)&imag);}
+  { int real = tex1Dfetch(tex1dchar4, array_index + (t)*NFREQUENCY*NSTATION*NPOL); \
+    int imag = tex1Dfetch(tex1dchar4, array_index + (NTIME_PIPE/4 + t)*NFREQUENCY*NSTATION*NPOL); \
+    CUBE_ADD_BYTES(4*sizeof(ComplexInput));				\
+    *(input##s##_p) = real;						\
+    *(input##s##_p + 4*TILE_WIDTH) = imag;}
 
 #else
-#error TEXTURE_DIM must be 1
+
+//#define TEXTURE_FLOAT_COORD
+#ifndef TEXTURE_FLOAT_COORD
+
+// Read float2 from global, write individual floats
+// to shared memory avoid bank conflict.
+#define LOAD(s, t)							\
+  {  int4 real, imag;							\
+    asm("tex.2d.v4.s32.s32 {%0, %1, %2, %3}, [tex2dchar4, {%4, %5}];" :	\
+	"=r"(real.x), "=r"(real.y), "=r"(real.z), "=r"(real.w) : "r"(array_index), "r"(t)); \
+    asm("tex.2d.v4.s32.s32 {%0, %1, %2, %3}, [tex2dchar4, {%4, %5}];" :	\
+	"=r"(imag.x), "=r"(imag.y), "=r"(imag.z), "=r"(imag.w) : "r"(array_index), "r"(t + NTIME_PIPE/4)); \
+    CUBE_ADD_BYTES(4*sizeof(ComplexInput));				\
+    *(input##s##_p) = real.x;						\
+    *(input##s##_p + 4*TILE_WIDTH) = imag.x;}
+
+#else
+
+// Read char4 from global, write individual floats
+// to shared memory avoid bank conflict.
+#define LOAD(s, t)							\
+  { int real = tex2D(tex2dchar4, array_index, t);			\
+    int imag = tex2D(tex2dchar4, array_index, t + NTIME_PIPE/4);	\
+    CUBE_ADD_BYTES(4*sizeof(ComplexInput));				\
+    *(input##s##_p) = *((int*)&real);					\
+    *(input##s##_p + 4*TILE_WIDTH) = *((int*)&imag);}
+#endif  // use float texture coordinates
+
 #endif
 
 // read in shared data as individual floats to avoid bank conflicts