--- a/src/xgpu.h
+++ b/src/xgpu.h
@@ -12,6 +12,10 @@ extern "C" {
 // point (i.e. floats).
 #define FIXED_POINT
 
+// If DP4A is defined, the library does all computation using native
+// 8-bit integer multiplication and 32-bit accumulation
+#define DP4A
+
 // set the data type accordingly
 #ifndef FIXED_POINT
 typedef float ReImInput;
@@ -24,10 +28,17 @@ typedef struct ComplexInputStruct {
   ReImInput imag;
 } ComplexInput;
 
+#ifndef DP4A
 typedef struct ComplexStruct {
   float real;
   float imag;
 } Complex;
+#else
+typedef struct ComplexStruct {
+  int real;
+  int imag;
+} Complex;
+#endif
 
 // Used to indicate the size and type of input data
 #define XGPU_INT8    (0)
@@ -66,6 +77,8 @@ typedef struct XGPUInfoStruct {
   unsigned int ntimepipe;
   // Type of input.  One of XGPU_INT8, XGPU_FLOAT32, XGPU_INT32.
   unsigned int input_type;
+  // Type of computation.  One of XGPU_INT8 or XGPU_FLOAT32
+  unsigned int compute_type;
   // Number of ComplexInput elements in input vector
   long long unsigned int vecLength;
   // Number of ComplexInput elements per transfer to GPU
@@ -216,6 +229,8 @@ void xgpuRandomComplex(ComplexInput* random_num, long long unsigned int length);
 void xgpuReorderMatrix(Complex *matrix);
 
 void xgpuCheckResult(Complex *gpu, Complex *cpu, int verbose, ComplexInput *array_h);
+
+void xgpuSwizzleInput(ComplexInput *out, const ComplexInput *in);
 
 void xgpuExtractMatrix(Complex *matrix, Complex *packed);