… src/xgpu_info.h
… src/xgpuinfo.c