This pull request adds support for using the `dp4a` instruction for native 8-bit multiply / 32-bit accumulate, present on all Pascal (excluding GP100 and GP10B) and Volta GPUs.
* This can accelerate xGPU by up to 4x versus the usual fp32 computation
* When using dp4a, the entire computation is pure integer, so the internal xGPU error checking is exact, not subject to rounding 
* Enabled with the `DP4A=yes` make command-line option
* Note that DP4A requires a different input data order, in particular the input data is swizzled with respect to time: `input[time/4][channel][station][pol][complexity][time%4]`.  At present xGPU offers no functionality to do this swizzling and expects the user to supply the data in this order.  The internal unit test does this swizzling on the host prior to copying to the GPU.
* Note that when using dp4a, there is the additional constraint that the `NTIME_PIPE` dimension must be a multiple of 16 (as opposed to 4).

Other changes in this pull request
* Default compilation set to Kepler sm_35 architecture since sm_20 has been removed in CUDA 9.x
* Remove `-Xptx abi=no` compilation option since this no longer exists with CUDA 9.x 
* Fix char sign ambiguity on POWER host processors


As an example of dp4a in action, below we show the performance 512 stations with and without dp4a on a Tesla V100 GPU.

**without dp4a**
```
$ ./bench CUDA_DIR=$CUDA_HOME NTIME=8192 NTIME_PIPE=1024 NSTATION=512 NFREQUENCY=10 CUDA_ARCH=sm_70 DP4A=no
===============================
CUBE Kernel profile
===============================

kernel                                  :    calls    time/s   Gflops Gflops/s  GiBytes GiBytes/s
cudaMemcpyHostToDevice_20971520         :        8  0.013626    0.000    0.000    0.156   11.467
shared2x2                               :        8  0.031117  344.268 11063.521    5.646  181.434
cudaMemcpyDeviceToHost_42024960         :        1  0.003229    0.000    0.000    0.039   12.123

Total                                   :       17  0.047972  344.268 7176.499    5.841  121.762

===============================
CUBE Asynchronous block profile
===============================

kernel                                  :    calls    time/s   Gflops Gflops/s  GiBytes GiBytes/s
ENTIRE_PIPELINE                         :        1  0.036123  344.268 9530.373    5.841  161.700
PIPELINE_LOOP                           :        1  0.027190  301.235 11078.789    5.077  186.712
```

**with dp4a**
```
$ ./bench CUDA_DIR=$CUDA_HOME NTIME=8192 NTIME_PIPE=1024 NSTATION=512 NFREQUENCY=10 CUDA_ARCH=sm_70 DP4A=yes
===============================
CUBE Kernel profile
===============================

kernel                                  :    calls    time/s   Gflops Gflops/s  GiBytes GiBytes/s
cudaMemcpyHostToDevice_20971520         :        8  0.013606    0.000    0.000    0.156   11.484
shared2x2                               :        8  0.007736  344.268 44500.285    5.646  729.772
cudaMemcpyDeviceToHost_42024960         :        1  0.003221    0.000    0.000    0.039   12.149

Total                                   :       17  0.024564  344.268 14015.052    5.841  237.791

===============================
CUBE Asynchronous block profile
===============================

kernel                                  :    calls    time/s   Gflops Gflops/s  GiBytes GiBytes/s
ENTIRE_PIPELINE                         :        1  0.017969  344.268 19158.621    5.841  325.061
PIPELINE_LOOP                           :        1  0.011969  301.235 25168.955    5.077  424.176
```