@maddyscientist 2018-05-30
Added table of GPU architectures showing the mapping between architecture, chip names and products to help set the compilation options.  Closes #13.