@maddyscientist 2018-05-30
@david-macmahon any objections to this?  I suggest also bumping the tag version to 2.0.0 once merged since the addition of dp4a is likely worthy of this.