@jack-h 2018-06-19
@david-macmahon @maddyscientist This seems to pass the benchmarks and fixes whatever issue I was having with commit 27e213af03

Thanks!