@maddyscientist 2018-06-21
Thanks @david-macmahon and @jack-h.  I have pushed a v2.0.0 tag as well now.